AUR link
Adds both the link to the AUR package, as well as a copy&paste command for the most popular AUR helper. At the time of writing 62,48 popularity compared to pikaur 9,16 popularity.

Could change it to a direct link to the package and omitting the copy&paste command, as suggested by "ignormies", because the most popular AUR helper changes frequently. But can also just make a new pull request in that case.