Fix sidebar length incorrect on resume
Prior to this change, the sidebar would persist the height it had when
the device entered sleep. If the device's height changed upon resume
(such as plugging in to an external display on a laptop) the sidebar
would remain at the old height, causing the sidebar to be too short or
too long for the display.

PapyElGringo/material-shell#123

Co-authored-by: Hilmar Gústafsson <hilmargustafs@gmail.com>