Activate AppPlaceholder only once, rename `clicked` property to `activated` (#649)
Previously, it was possible through an impatient click while waiting for a
slow app to start to activate an app multiple times. That results in multiple
emissions of the `request-new-meta-window` events and chaos ensues.

Fixes #649 by checking and setting `waitForReset` to `true` after a call to
`activate` and emitting `activated` only as the very last action of that
function.