Make 'tileableList-changed' not send the old list with the event.
Previously theoretical race conditions could cause the old list to be inaccurate (if multiple updates happened before the event fired).
Now the receivers that need to diff against the old list will keep track of this info themselves, which should be more robust.
This also simplifies the code a bit.