Fix wrong animation direction in split layout. (#824)
When you had more than 3 tileables in a split layout and
changed focus in a way that completely swapped out the tileables on the screen,
then the split layout would previously always animate the windows going in one direction.

This changes it to set the animation direction more accurately even if the set of tileablesis completely swapped out.