Support configurable install location and bare make
My motivation for this is to support NixOS which installs into a
content-addressed directory layout and needs to be able to provide the
location of the XDG share (specific to the package)

Also it is more conventional to have the first target in make file do
the 'full build', then the standard build of `make && make install` will
work. I couldn't see a downside of adding this so I added an `all`
target. For my purposes this makes the packaging for NixOS more minimal
since I don't need to customise the standard builder (which is easy
enough to do, but this seemed like a good thing to have anyway).

Signed-off-by: Silas Davis <silas.davis@monax.io>