Support manually overriding the git commit hash for "metadata.json".
The default for the "commit" field in "metadata.json" continues to be the git commit hash of the currently checked out git revision.
If `make compile` is run in a directory created without git, e.g. by extracting an archive, the "commit" field has, by default, an empty string as a value.
Both defaults can be overwritten with the value of the environment variable "GIT_COMMIT_HASH".
This allows distro packagers to have an archive as the source and still provide the commit revision in the usual build process.