Implement a recent search provider for the sidebar search.
Since the providers have previously been ordered sequentially, you often
have to either use the mouse or press 'down' many times to get to the item that you want,
even if you open the exact same app every time.

This commit records the history of which things you actually open and suggests those when relevant.
It will only suggest a single item at a time and uses various pieces of information such as what the user is searching for, what applications are open, the current time of day, and a few more.

If it cannot decide between several items in the history, then nothing will be shown. Most likely a suggestion will show up if the user types a few more characters.